+++ src/main/java/com/wagglex2/waggle/domain/auth/service/AuthService.java
@@ -1,8 +1,11 @@
 package com.wagglex2.waggle.domain.auth.service;
 
-public interface AuthService {
-    public void sendAuthCode(String toEmail);
-    public void sendEmailAuthCode(String toEmail, String verificationCode);
-    public void verifyCode(String toEmail, String inputCode);
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
+public interface AuthService {
+    void sendAuthCode(String toEmail);
+    void sendEmailAuthCode(String toEmail, String verificationCode);
+    void verifyCode(String toEmail, String inputCode);
+    void issueNewTokens(HttpServletRequest request, HttpServletResponse response);
 }
